extract material slots in data grid pro
