+++ packages/grid/x-data-grid-pro/src/components/index.ts
@@ -2,4 +2,4 @@
 export * from './GridTreeDataGroupingCell';
 export * from './GridColumnMenuPinningItem';
 export * from './GridDetailPanelToggleCell';
-export * from './icons';
+export * from '../material/icons';
